!!!WARNING!!! - INDEX FORMAT CHANGED.
I've changed the index format slightly so that I can add compression
functionality (still to be added). It may change even further in the next
couple of revisions. Please only update if you want to experiment with the
latest version. Not to be used in production. I apologize for any
inconvenience.

As a side note, document loading should be even faster.


git-svn-id: svn+ssh://davebalmain.com/home/dave/repos/ferret/trunk@852 06fd6eb0-0002-0410-a719-e5602cce40bc