Ferret will use bzlib instead of zlib as the license is compatible and the compression is better.
Although bzlib is much slower than zlib, it won't be that noticable as there is
already going to be a lot of processing happening for each field in the index.
Also, if you really want to compress your index data, you may as well go for
the best compression you can get.


git-svn-id: svn+ssh://davebalmain.com/home/dave/repos/ferret/trunk@855 06fd6eb0-0002-0410-a719-e5602cce40bc