Added .gitignore for ruby directory.
The ruby bindings statically link most of the files from the C source for
compilation. We don't want to store these twice in the git repository. Although
it doesn't make any difference in git, these changes also get pushed to an svn
repository so it is important that these files don't get included twice.
Instead they are linked (or copied in the case of Windows) during the build of
the binaries or the gem.


git-svn-id: svn+ssh://davebalmain.com/home/dave/repos/ferret/trunk@921 06fd6eb0-0002-0410-a719-e5602cce40bc