Added rake show:badnames[file] task. Cleaned bitvector.h.
I added a couple of tasks to print the symbol table for files. The interesting
one is rake show:badnames which scans for public names that do not have the
(frt_, FRT_, Frt) prefixes.

I also changed the default prefix for structs form Ferret to Frt to be the same
as constants and functions. It adds consistency and saves some characters. Some
of the type names just looked too long to me (like FerretPriorityQueue for
example).

Also cleaned the public symbol table for bitvector.h


git-svn-id: svn+ssh://davebalmain.com/home/dave/repos/ferret/trunk@924 06fd6eb0-0002-0410-a719-e5602cce40bc