Finished cleaning up public namespace. However namespace will still be in flux.
I've cleaned up the public namespace so that all public names are prefixed with
a frt_, FRT_ or Frt (for functions, constants and types respectivly). However,
I'm still not happy with it and I'd like to do a little more work on it. For
starters there are a lot of types and functions that have been made globel so
that they can be tested. I'm considering modifying the unit tests so that you
can include the unit tests with the code. These tests could be surrounded with
a "#define UNIT_TESTS" or something so they won't be included in the final
product. I'd also add a rake task which searches for unit tests and
automatically adds them to the test suite so they get run. I haven't tried this
yet so I'm still not sure how it will work.


git-svn-id: svn+ssh://davebalmain.com/home/dave/repos/ferret/trunk@941 06fd6eb0-0002-0410-a719-e5602cce40bc