Fixed stacktrace tmpnam issue. Now uses .gdb-bt
Instead of using tmpnam and creating a temporary file within C I just added an
`echo "bt\nq\n" > .gdb-bt &&` to the start of the gdb command. This seemed to
be the simplest way to solve the problem.

Also added -DHAVE_GDB define to the Rakefile automatically if gdb is present.
Make will no longer use the gdb stacktrace by default. We can add this later if
we decide to use autotools.


git-svn-id: svn+ssh://davebalmain.com/home/dave/repos/ferret/trunk@954 06fd6eb0-0002-0410-a719-e5602cce40bc