Automated adding of tests.
Just declare each TestSuite like so;

    TestSuite *ts_<name of test suite>(TestSuite *suite)

A rake task will automatically search the test files for new test suites and
add them to the list to be run.

Also removed the bench executables from the clean-task lists in the Makefiles.
These are going to be moved to the bench suite.


git-svn-id: svn+ssh://davebalmain.com/home/dave/repos/ferret/trunk@959 06fd6eb0-0002-0410-a719-e5602cce40bc