Release of ragel-based standard tokenizer
* Removed all of the old tokenizer code from analysis.c
* Changed the FrtStandardTokenizer struct to include a 'is_ascii' member
  and got rid of the function pointers
* Ascii and multibyte tokenizers pass all unit tests
* They are currently in 2 separate Ragel files, i'm looking into merging them
  since the differences are minimal.  Maybe I can use the C preprocessor
  to generate both versions from the same state machine.

NOTE: this is actually a resync with the first git repository so there are
extra changes in here. I'm resyncing the repository because of the mess created
while trying to keep the git repo in sync with the svn repo. There seemed to be
some recurring problems so I started again from scratch.


git-svn-id: svn+ssh://davebalmain.com/home/dave/repos/ferret/trunk@981 06fd6eb0-0002-0410-a719-e5602cce40bc