Updated Rakefile to work with multibyte ragel scanner
Rakefile will now automatically detect new Ragel files if the exist and compile
them. To implement this, ragel and yacc files are now compiled with rules.

Also made a slight modification to scanner when scanning hyphenated words.

-        walnum+ ('-' walnum+)* { RET; };
+        walnum+ ([\-_] walnum+)* { RET; };

This change is simply to get the ruby unit tests to pass. Still need to
investigate the full difference between this scanner and the previous scanner.


git-svn-id: svn+ssh://davebalmain.com/home/dave/repos/ferret/trunk@989 06fd6eb0-0002-0410-a719-e5602cce40bc