Added vint writing benchmark to test loop unrolling
Turns out that we need to work out exactly how many bytes we have to write
whichever way we look at it so I still haven't thought of a way to improve this
methods. I assume reading vints would be the same. These methods are important
as they are basically the bottle neck for all operations on the index.

Also added -D_BSD_SOURCE to get rid of error about not declaring mkstemp. This
error shouldn't be a problem on windows as mkstemp is only used in the HAVE_GDB
code which will presumably not be used on windows.


git-svn-id: svn+ssh://davebalmain.com/home/dave/repos/ferret/trunk@1006 06fd6eb0-0002-0410-a719-e5602cce40bc