Made rake recompile when compile flags change
You can now call rake test followed by rake benchmark++ and rake will recompile
everything with the new flags as required. No need to run rake clean.

Also note the addition of the ++ methods which compile with c++ instead of c.
This currently done simply for the extra type checking and so that we can
benchmark some of the core data structures against those found in the c++ stl.


git-svn-id: svn+ssh://davebalmain.com/home/dave/repos/ferret/trunk@1009 06fd6eb0-0002-0410-a719-e5602cce40bc