Fixed ruby unit tests. Tests currently fail due to ragel bug.
There seems to be a bug in ragel which is causing tests to fail. An example of
the code to reproduce the problem can be found at;

    http://pastie.caboo.se/183409

Also fixed a problem in the ruby unit tests. Text that was not matching with
the previous standard tokenizer now matches. Since this isn't a bug I slightly
modified the unit tests.

Note that both the ruby and C unit tests currently fail. I hope to have the
resolved as soon as possible.


git-svn-id: svn+ssh://davebalmain.com/home/dave/repos/ferret/trunk@1017 06fd6eb0-0002-0410-a719-e5602cce40bc