Apoligies everyone for the problems with the repository
Sorry, about the constantly changing subversion repository. We recently started
working in git and I ran into some major issues with syncing the git repository
with subversion. The problems seem to be behind us now so hopefully everything
should remain stable from now on.

The only thing that changed in this commit were a few additions to the TODO list


git-svn-id: svn+ssh://davebalmain.com/home/dave/repos/ferret/trunk@1025 06fd6eb0-0002-0410-a719-e5602cce40bc