Fixed ragel big in standard tokenizer
The Ragel fix didn't actually work but I managed to find a hack that fixes the
problem. Basically I just add an empty action to the "junk skipper" at the end.

-        (any - frt_alnum);
+        (any - frt_alnum) {};

So now all the tests pass. I'm in the process of working on a patch for Adrian
to get this problem fixed in ragel. I also noticed one other difference in the
latest version of ragel. The include paths seem to now be relative to the
including file rather than the calling process to I removed the "src/" from all
the include calls, so you'll still need to install the latest ragel. Hopefully
the fix will be checked in by the time you do.