Added the intern method. All field names are now interned.
Whenever a field name is assigned to a struct it must be interned. This means that field names never need to be freed and they can be compared with == instead of strcmp. This is not the case for public functions which have a field parameter. In this case, the parameter will either need to be interned or the standard string methods used for comparison.

I'm not sure I like this solution yet. The reason I've added it is that it means there is one less object that we need to keep track of in terms of memory allocation. It also adds a slight speed and memory usage improvement as field names only ever get allocated once and they are quite commonly compared so it is more efficient to use == instead of strcmp. However, it feels like it could be quite easy to forget to intern a field somewhere causing a hard to track down error. In future I may add a Symbol object to force field names to be interned.


git-svn-id: svn+ssh://davebalmain.com/home/dave/repos/ferret/trunk@1027 06fd6eb0-0002-0410-a719-e5602cce40bc