Started detailed documentation of the QueryParser
In writing the documentation I've spotted a pretty serious bug in the
QueryParser. Fields need to be pushed onto a stack as they are specified,
otherwise the parser is quite broken for multi-level, multi-field queries. I'll
fix this first thing tomorrow before finishing this documentation. Comments on
the documentation are welcome.