Fixed field specifier bug in QueryParser
The QueryParser now uses a stack to keep track of assigned fields in queries.
This means that the query;

    f1:(aaa f2:bbb ccc)

will parse correctly as;

    f1:aaa f2:bbb f1:ccc

Previously the ccc was either restricted to the default fields or got dropped
from the query all together.

PS: I've already checked this in and tried to dcommit it again and all of a
sudden the commit seems to have disappeared. I'm getting pretty tired of
git-svn.


git-svn-id: svn+ssh://davebalmain.com/home/dave/repos/ferret/trunk@1032 06fd6eb0-0002-0410-a719-e5602cce40bc