Modifed integer Hash to also accept pointers
Integer hash has been improved so that it stores a dummy key instead of any old
junk which it was storing before. The h_lookup_int has been changed to
h_lookup_ptr but it is still used by h_set_int etc. I'll soon be adding
h_new_ptr which will be an optimized hash, especially for storing Symbols.