Changed ruby bindings to use ruby symbols
Ruby bindings now replace Ferret's Symbol implementation with its own. The main
advantage of this is that we don't need to do as much to convert Ferret's data
types into Ruby data types.

An added bonus of this was that it highlighted all the places where the current
Ferret Symbol implementation was broken so it is much more solid now.