Added utf-8 standard analyzer. Back up to speed.
Added utf-8 standard analyzer which is on par with the legacy standard
analyzer. I'm still not happy with this solution though. I think the best idea
right now is to write our own generator which can make use of C's standard
multibyte functions. This shouldn't be too hard as we aren't using many of
Ragel's advanced features anyway.